35 change UI refactor to use tamagui (#36)
Large change.

Merges 33# ( Zustand addition )

This brings in Tamagui as the new UI library, sets up base configuration for it and refactors landing page and bottom bar to use tamagui. These refactors are NOT perfect, but they are provided to help understand how Tamagui works.

What needs to be done in the future:

    Refactor other pages with Tamagui
    Clean up and finalize refactor of landing page
    ( things like removing constant size definitions from landing and bottom bar )
    Complete config to make our theme fully realized, and fit for use.